+++ brion/nodeGroup.h
@@ -40,14 +40,17 @@ class NodeGroup
 {
 public:
     friend class Nodes;
+    NodeGroup();
     NodeGroup(NodeGroup&&);
     ~NodeGroup();
 
+    NodeGroup& operator=(NodeGroup&&);
+
     /** Return the names of all the datasets of the node group */
-    Strings getAttributeNames();
+    Strings getAttributeNames() const;
 
     /** Return the names of all the datasets in the 'dynamics_params' group */
-    Strings getDynamicParameterNames();
+    Strings getDynamicParameterNames() const;
 
     /** Return the number of nodes in the node group */
     size_t getNumberOfNodes() const;